+++ src/widget/scrollable/mod.rs
@@ -1,0 +1,6 @@
+// Copyright 2022 System76 <info@system76.com>
+// SPDX-License-Identifier: MPL-2.0
+
+mod scrollable;
+
+pub use scrollable::{horizontal, scrollable, vertical};
